# public/api-docs.yaml
openapi: 3.0.3
info:
  title: Assessment API - Video Platform
  description: A comprehensive video sharing platform API with authentication, video management, social features, and content discovery
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
servers:
  - url: http://localhost:8000/api
    description: Development server
security:
  - JWT: []
components:
  securitySchemes:
    JWT:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for authentication
...
paths:
...
  /videos/user/{userId}:
    get:
      tags:
        - Videos
      summary: Get user's videos
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: User videos retrieved

  # Video Interactions
...
  /profile/{id}:
    get:
      tags:
        - User Profile
      summary: Get user profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Profile retrieved
  /profile:
    put:
      tags:
        - User Profile
      summary: Update profile
      security:
        - JWT: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                firstName:
                  type: string
                bio:
                  type: string
                profilePicture:
                  type: string
      responses:
        '200':
          description: Profile updated successfully
  /profile/{id}/follow:
    post:
      tags:
        - User Profile
      summary: Follow user
      security:
        - JWT: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: User followed successfully
  /profile/{id}/unfollow:
    delete:
      tags:
        - User Profile
      summary: Unfollow user
      security:
        - JWT: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: User unfollowed successfully
  /profile/{id}/followers:
    get:
      tags:
        - User Profile
      summary: Get user followers
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: page
          in: query
          schema:
            type: integer
            default: 1
      responses:
        '200':
          description: Followers retrieved
  /profile/{id}/following:
    get:
      tags:
        - User Profile
      summary: Get user following
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: page
          in: query
          schema:
            type: integer
            default: 1
      responses:
        '200':
          description: Following retrieved
  /profile/search:
    get:
      tags:
        - User Profile
      summary: Search users
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
            minLength: 2
        - name: page
          in: query
          schema:
            type: integer
            default: 1
      responses:
        '200':
          description: Users found
...